Stash progress (amend this)
NOTE: currently the columns can be seen in the attribute table (and
the fields tab in layer properties). The values however are not shown,
and that was the next step I was trying to do. Currently I was working
only with the temporal ranges, but it might be better to just include
all of them at the same time.